Add theme reducer
